chore: put the redis subscriber in its own file for later testing
